Extract KewlEntity* from MigrationTests
This is in preparation for a new MigrationRunner which will also need to use them.
