+++ test/functional/Migration/MigrationTests.cs
@@ -3,8 +3,6 @@ using System.Reflection;
 using FakeItEasy;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
-using MongoDB.Bson;
-using MongoDB.Bson.Serialization.Attributes;
 using RapidCore.Locking;
 using RapidCore.Mongo.FunctionalTests.Migration.TestMigration;
 using RapidCore.Mongo.Migration;
@@ -65,28 +63,6 @@ namespace RapidCore.Mongo.FunctionalTests.Migration
             var sevenUp = await db.FirstOrDefaultAsync<KewlEntityUpdated>(KewlEntity.Collection, x => x.Id == seven.Id);
             Assert.Equal("7", sevenUp.Reference);
             Assert.Equal("Bubbly", sevenUp.Mucho);
-        }
-        
-        public class KewlEntity
-        {
-            [BsonIgnoreIfDefault]
-            public ObjectId Id { get; set; }
-            
-            public int Reference { get; set; }
-
-            public static string Collection => "Kewl";
-        }
-        
-        public class KewlEntityUpdated
-        {
-            [BsonIgnoreIfDefault]
-            public ObjectId Id { get; set; }
-            
-            public string Reference { get; set; }
-            
-            public string Mucho { get; set; }
-            
-            public static string Collection => "Kewl";
         }
     }
 }
