+++ test/functional/Migration/TestMigration/Migration02.cs
@@ -12,7 +12,7 @@ namespace RapidCore.Mongo.FunctionalTests.Migration.TestMigration
             
             builder.Step("Change Reference from int to string", () =>
             {
-                var collection = db.GetCollection<MigrationTests.KewlEntity>(MigrationTests.KewlEntity.Collection);
+                var collection = db.GetCollection<KewlEntity>(KewlEntity.Collection);
 
                 var cursor = collection
                     .Find(_ => true)
@@ -20,10 +20,10 @@ namespace RapidCore.Mongo.FunctionalTests.Migration.TestMigration
 
                 return cursor.ForEachAsync(doc =>
                 {
-                    var filter = Builders<MigrationTests.KewlEntityUpdated>.Filter.Eq(x => x.Id, doc.Id);
-                    var update = Builders<MigrationTests.KewlEntityUpdated>.Update.Set("Reference", doc.Reference.ToString());
+                    var filter = Builders<KewlEntityUpdated>.Filter.Eq(x => x.Id, doc.Id);
+                    var update = Builders<KewlEntityUpdated>.Update.Set("Reference", doc.Reference.ToString());
 
-                    db.GetCollection<MigrationTests.KewlEntityUpdated>(MigrationTests.KewlEntityUpdated.Collection).UpdateMany(filter, update);
+                    db.GetCollection<KewlEntityUpdated>(KewlEntityUpdated.Collection).UpdateMany(filter, update);
                 });
             });
         }
